refactors functions back into Beef.Users module
