+++ kousa/lib/beef/mutations/users.ex
@@ -1,99 +1,91 @@
-defmodule Beef.Queries.User do
-
+defmodule Beef.Mutations.Users do
   import Ecto.Query, warn: false
 
   alias Beef.Repo
-  alias Beef.Rooms
-  alias Beef.Schemas.Room
   alias Beef.Schemas.User
 
-  @fetch_limit 16
-  def search(query, offset) do
-    query_with_percent = "%" <> query <> "%"
-
-    items =
-      from(u in User,
-        where:
-          ilike(u.username, ^query_with_percent) or
-            ilike(u.displayName, ^query_with_percent),
-        left_join: cr in Room,
-        on: u.currentRoomId == cr.id and cr.isPrivate == false,
-        select: %{u | currentRoom: cr},
-        limit: @fetch_limit,
-        offset: ^offset
-      )
-      |> Repo.all()
-
-    {Enum.slice(items, 0, -1 + @fetch_limit),
-     if(length(items) == @fetch_limit, do: -1 + offset + @fetch_limit, else: nil)}
+  def edit_profile(user_id, data) do
+    %User{id: user_id}
+    |> User.edit_changeset(data)
+    |> Repo.update()
   end
 
-  def get_users_in_current_room(user_id) do
-    case tuple_get_current_room_id(user_id) do
-      {:ok, current_room_id} ->
-        {current_room_id,
-         from(u in User,
-           where: u.currentRoomId == ^current_room_id,
-           left_join: rp in Beef.RoomPermission,
-           on: rp.userId == u.id and rp.roomId == u.currentRoomId,
-           select: %{u | roomPermissions: rp}
-         )
-         |> Repo.all()}
-
-      _ ->
-        {nil, []}
-    end
+  def delete(user_id) do
+    %User{id: user_id} |> Repo.delete()
   end
 
-  def tuple_get_current_room_id(user_id) do
-    case Kousa.RegUtils.lookup_and_call(
-           Kousa.Gen.UserSession,
-           user_id,
-           {:get_current_room_id}
-         ) do
-      {:ok, nil} ->
-        {nil, nil}
-
-      x ->
-        x
-    end
+  def bulk_insert(users) do
+    Repo.insert_all(
+      User,
+      users,
+      on_conflict: :nothing
+    )
   end
 
-  @spec get_by_id_with_current_room(any) :: any
-  def get_by_id_with_current_room(user_id) do
+  def inc_num_following(user_id, n) do
     from(u in User,
-      left_join: a0 in assoc(u, :currentRoom),
       where: u.id == ^user_id,
-      limit: 1,
-      preload: [
-        currentRoom: a0
+      update: [
+        inc: [
+          numFollowing: ^n
+        ]
       ]
     )
-    |> Repo.one()
+    |> Repo.update_all([])
   end
 
-  def get_current_room(user_id) do
-    room_id = get_current_room_id(user_id)
-
-    case room_id do
-      nil -> nil
-      id -> Rooms.get_room_by_id(id)
-    end
+  def set_reason_for_ban(user_id, reason_for_ban) do
+    from(u in User,
+      where: u.id == ^user_id,
+      update: [
+        set: [
+          reasonForBan: ^reason_for_ban
+        ]
+      ]
+    )
+    |> Repo.update_all([])
   end
 
-  def get_current_room_id(user_id) do
-    case Kousa.RegUtils.lookup_and_call(
-           Kousa.Gen.UserSession,
-           user_id,
-           {:get_current_room_id}
-         ) do
-      {:ok, id} ->
-        id
-
-      _ ->
-        nil
-    end
+  def set_online(user_id) do
+    from(u in User,
+      where: u.id == ^user_id,
+      update: [
+        set: [
+          online: true
+        ]
+      ]
+    )
+    |> Repo.update_all([])
   end
+
+  def set_user_left_current_room(user_id) do
+    Kousa.RegUtils.lookup_and_cast(Kousa.Gen.UserSession, user_id, {:set_current_room_id, nil})
+
+    from(u in User,
+      where: u.id == ^user_id,
+      update: [
+        set: [
+          currentRoomId: nil
+        ]
+      ]
+    )
+    |> Repo.update_all([])
+  end
+
+  def set_offline(user_id) do
+    from(u in User,
+      where: u.id == ^user_id,
+      update: [
+        set: [
+          online: false,
+          lastOnline: fragment("now()")
+        ]
+      ]
+      # select: u
+    )
+    |> Repo.update_all([])
+  end
+
 
   def set_current_room(user_id, room_id, can_speak \\ false, returning \\ false) do
     roomPermissions =
