+++ kousa/lib/beef/queries/users.ex
@@ -1,0 +1,6 @@
+defmodule Beef.Queries.User do
+  @moduledoc """
+  all functions in this module should be "Query builder" functions,
+  they should not touch the database.
+  """
+end
